Move plot to plots
