+++ requirements.txt
@@ -8,8 +8,7 @@ scikit-learn
 PyYAML==6.0
 jenkspy
 numpy
-# umap-learn==0.5.3 # removed due to lack of space
-# tensorflow==2.14.0
+umap-learn==0.5.3
 dash
 dash-bootstrap-components
 nbformat>=4.2.0
@@ -18,3 +17,4 @@ rdkit
 pyarrow
 jinja2
 dash_dangerously_set_inner_html
+hdbscan
